requirements.txt
fastapi==0.104.1
rasterio==1.3.9
Pillow==10.1.0
numpy==1.26.2
uvicorn==0.24.0
python-multipart==0.0.6
